+++ src/common/Smi.Common/Messages/MessageHeader.cs
@@ -12,8 +12,6 @@ namespace Smi.Common.Messages
 {
     public class MessageHeader : IMessageHeader, IEquatable<MessageHeader>
     {
-        // FIXME: Pull out header property strings into const. variables
-
         public Guid MessageGuid { get; set; }
 
         public int ProducerProcessID { get; set; }
@@ -180,7 +178,9 @@ namespace Smi.Common.Messages
                 hashCode = (hashCode * 397) ^ ProducerProcessID;
                 hashCode = (hashCode * 397) ^ (ProducerExecutableName != null ? ProducerExecutableName.GetHashCode() : 0);
                 hashCode = (hashCode * 397) ^ OriginalPublishTimestamp.GetHashCode();
-                hashCode = (hashCode * 397) ^ (Parents != null ? Parents.GetHashCode() : 0);
+                // NOTE(rkm 2020-03-04) GetHashCode for a struct[] uses reference equality, so instead we compute the hash
+                //                      code using the string representation of the Parents array
+                hashCode = (hashCode * 397) ^ (Parents != null ? string.Join(Splitter, Parents).GetHashCode() : 0);
                 return hashCode;
             }
         }